+++ src/main/java/hu/uni/eku/tzs/service/DirectorManagerImpl.java
@@ -1,12 +1,10 @@
 package hu.uni.eku.tzs.service;
 
 import hu.uni.eku.tzs.dao.DirectorRepository;
-import hu.uni.eku.tzs.dao.entity.ActorEntity;
 import hu.uni.eku.tzs.dao.entity.DirectorEntity;
 import hu.uni.eku.tzs.model.Director;
-import hu.uni.eku.tzs.service.exceptions.ActorNotFoundException;
-import hu.uni.eku.tzs.service.exceptions.DirectorAlreadyExistsException;
 import hu.uni.eku.tzs.service.exceptions.DirectorNotFoundException;
+import hu.uni.eku.tzs.service.exceptions.DirectorAlreadyExistsException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -20,34 +18,34 @@ public class DirectorManagerImpl implements DirectorManager {
 
     private final DirectorRepository directorRepository;
 
-    private static Director convertDirectorEntityToModel(DirectorEntity directorEntity){
+    private static Director convertDirectorEntityToModel(DirectorEntity directorEntity) {
         return new Director(
-                directorEntity.getDirectorId(),
-                directorEntity.getDirectorQuality(),
-                directorEntity.getAverageRevenue()
+            directorEntity.getDirectorId(),
+            directorEntity.getDirectorQuality(),
+            directorEntity.getAverageRevenue()
         );
     }
 
-    private static DirectorEntity convertDirectorModelToEntity(Director director){
+    private static DirectorEntity convertDirectorModelToEntity(Director director) {
         return DirectorEntity.builder()
-                .directorId(director.getDirectorId())
-                .directorQuality(director.getDirectorQuality())
-                .averageRevenue(director.getAverageR())
-                .build();
+            .directorId(director.getDirectorId())
+            .directorQuality(director.getDirectorQuality())
+            .averageRevenue(director.getAverageR())
+            .build();
     }
 
     @Override
-    public Director record(Director director) throws DirectorNotFoundException, DirectorAlreadyExistsException {
-        if(directorRepository.findById(director.getDirectorId()).isPresent()){
+    public Director record(Director director) throws DirectorAlreadyExistsException {
+        if (directorRepository.findById(director.getDirectorId()).isPresent()) {
             throw new DirectorAlreadyExistsException();
         }
 
         DirectorEntity directorEntity = directorRepository.save(
-                DirectorEntity.builder()
-                        .directorId(director.getDirectorId())
-                        .directorQuality(director.getDirectorQuality())
-                        .averageRevenue(director.getAverageR())
-                        .build()
+            DirectorEntity.builder()
+                .directorId(director.getDirectorId())
+                .directorQuality(director.getDirectorQuality())
+                .averageRevenue(director.getAverageR())
+                .build()
         );
         return convertDirectorEntityToModel(directorEntity);
     }
@@ -55,7 +53,7 @@ public class DirectorManagerImpl implements DirectorManager {
     @Override
     public Director readById(int id) throws DirectorNotFoundException {
         Optional<DirectorEntity> entity = directorRepository.findById(id);
-        if(entity.isEmpty()){
+        if (entity.isEmpty()) {
             throw new DirectorNotFoundException(String.format("Cannot find a director with this ID: %s", id));
         }
         return convertDirectorEntityToModel(entity.get());
@@ -64,17 +62,23 @@ public class DirectorManagerImpl implements DirectorManager {
     @Override
     public Collection<Director> readAll() {
         return directorRepository.findAll().stream().map(DirectorManagerImpl::convertDirectorEntityToModel)
-                .collect(Collectors.toList());
+            .collect(Collectors.toList());
     }
 
     @Override
     public Director modify(Director director) throws DirectorNotFoundException {
         DirectorEntity entity = convertDirectorModelToEntity(director);
+        if (directorRepository.findById(director.getDirectorId()).isEmpty()) {
+            throw new DirectorNotFoundException("Can not modify director with this id");
+        }
         return convertDirectorEntityToModel(directorRepository.save(entity));
     }
 
     @Override
     public void delete(Director director) throws DirectorNotFoundException {
+        if (directorRepository.findById(director.getDirectorId()).isEmpty()) {
+            throw new DirectorNotFoundException("Can not delete a director with this id");
+        }
         directorRepository.delete(convertDirectorModelToEntity(director));
     }
 }
